+++ src/noahtheduke/splint.clj
@@ -45,6 +45,7 @@
 
     noahtheduke.splint.rules.naming.conventional-aliases
     noahtheduke.splint.rules.naming.conversion-functions
+    noahtheduke.splint.rules.naming.lisp-case
     noahtheduke.splint.rules.naming.predicate
     noahtheduke.splint.rules.naming.single-segment-namespace
     noahtheduke.splint.rules.naming.record-name